feat: support for include location in server {} block #48
